Refactor the direction calculation to account for early food
The way we were calculating the direction didn't account for initial
conditions where body parts overlap. This caused the tail to be rendered
in the incorrect direction. As a fix, we looked at the partIndex - 2,
which was equivalently the head of the snake, when determining the
direction.

This works, except when a snake eats food in the first turn. Now an
index offset of -2 doesn't refer to the head of the snake.

As a refactor, use the unique coordinates occupied by a snake for the
direction calculation. As an example: even if a snake has a body length
of 3, after the first move it will only occupy two spots on the board.
And so we treat it as a snake of length 2 when calculating direction.