Integrate react router (#39)
Install and set up initial router to allow navigating to board settings page. This is the initial plumbing with a placeholder settings page. The default behaviour of the board should be unchanged.

I also snuck in a cleanup task and upgraded react version from 16.8.4 to 17.0.2 (latest stable). This required updating related packages as well as removing some that we were no longer using, or were made redundant by the updated packages.

Changelogs can be reviewed here:

* https://github.com/facebook/react/blob/main/CHANGELOG.md#16100-september-27-2019
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16101-september-28-2019
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16102-october-3-2019
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16110-october-22-2019
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16120-november-14-2019
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16130-february-26-2020
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16131-march-19-2020
* https://github.com/facebook/react/blob/main/CHANGELOG.md#16140-october-14-2020
* https://github.com/facebook/react/blob/main/CHANGELOG.md#1700-october-20-2020
* https://github.com/facebook/react/blob/main/CHANGELOG.md#1702-march-22-2021
* https://github.com/reduxjs/react-redux/releases/tag/v7.2.4
* https://github.com/reduxjs/react-redux/releases/tag/v7.2.5