Playback speed component (#40)
Initial implementation of persisted playback speed settings. It should allow setting the value by navigating to /settings. This will persist the value in localStorage (if available). When navigating back to the board / the persisted value is used as the framerate value already in use by the code and referenced by the docs.