Change settings speeds (#48)
The slow speed isn't feeling quite right so we're adjusting the values:
* SLOW: 3fps (old 6fps)
* MEDIUM: 6fps (old 10fps)
* FAST: 20 fps (no change)
This change is not fully backwards compatible since any previously saved slow speed will now display as medium in the settings form. It won't affect the playback though.

To add more confidence in our components a new test file was added for the PlaybackSpeed component.
Tests verify that the correct framerate is sent when a new speed is selected in the settings form AND the slider will display the correct speed using a range of values. This will improve backwards compatibility with previous settings that are saved to localStorage.
* SLOW: 1 - 5 (previous 6)
* MEDIUM: 6 - 12 (previous 10)
* FAST: 13 - 20 (previous 20)

Also includes some cleanup. The default settings values are now defined in one place and imported where needed.