Add Coordinate Labels
First we adjust the size of the `Grid` we are going to render within `Board`

If we have the `showCoordinateLabels` option enabled, make the `Grid` inside
smaller, so that we have room for the labels.

The labels are rendered inside `Grid` but OUTSIDE it's SVGs bounding box.
We allow the SVG to overflow so that the labels are visible, when the option
is set.
This allowed us to use `-1` for the row/col where the labels went and means
we didn't need to modify the rest of the grid generation for this PR.

To get the text centered nicely in the cell, we use a `foreignObject`
and HTML + CSS. SVGs don't seem to support text styling as well as HTML,
and foreignObject seems supported in all non-IE browsers