Fix Deploys: Move all Logic from `load` to `onMount` (#80)
Builds are currently failing with this error message

Accessing url.searchParams during prerendering is forbidden. If you need to use it, ensure you are only doing so in the browser (for example in onMount).

It was introduced in #79

This error is because we are trying to access url.searchParams in the load function and ALSO have preloading enabled.
Since load gets called on the server to pre-render, it doesn't allow us to access url.searchParams since that can only be known in the browser

But this must be new behavior, since it was working for us previously. I think it was working previously is because load was acting as a 'universal` loader and running both in the 'server' when preloading, and also in the client when mounting for the first time.

kit.svelte.dev/docs/load#universal-vs-server-when-does-which-load-function-run

I imagine the server load was getting an empty searchSet, but continuing to work. And then the client load correctly filled in the query param values

But from reading the Changelog I can't find anything that seems obviously like it would have broken this behavior

I do think this is a good change to prepare for v2, since it seems required there as well!